src/requirements.txt
numpy
pandas
imageio
pillow
seaborn
matplotlib
mujoco-py==v1.50.1.0
tqdm
torch>=1.1.0
gpytorch==0.3.5
tensorboardX==1.9
gym